Bug 1138870 - [Stingray] Move content of tv_shared folder to shared folder
