+++ shared/js/smart-screen/animations.js
