+++ shared/js/smart-screen/applications.js
