+++ shared/js/smart-screen/async_semaphore.js
