+++ shared/js/smart-screen/card_store.js
