+++ shared/js/smart-screen/cards/application.js
