+++ shared/js/smart-screen/cards/card.js
