+++ shared/js/smart-screen/clock.js
