+++ shared/js/smart-screen/init_gesture.js
