+++ shared/js/smart-screen/key_navigation_adapter.js
