+++ shared/js/smart-screen/piped_promise.js
