+++ shared/js/smart-screen/selection_border.js
