+++ shared/js/smart-screen/settings_cache.js
