+++ shared/js/smart-screen/shared_utils.js
