+++ shared/js/smart-screen/simple_key_navigation.js
