+++ shared/js/smart-screen/spatial_navigator.js
