+++ shared/resources/default-cards.json
