+++ shared/style/smart-screen/animations.css
