+++ shared/test/unit/mocks/mock_piped_promise.js
