+++ shared/test/unit/mocks/mock_presentation.js
