+++ shared/test/unit/mocks/mock_presentation_session.js
