+++ shared/test/unit/mocks/mock_xml_http_request.js
