+++ tv_apps/app-deck/style/icons/ic_delete.png
