+++ tv_apps/app-deck/style/icons/ic_pin.png
