+++ tv_apps/app-deck/style/icons/ic_pin@2x.png
