+++ tv_apps/fling-player/index.html
@@ -6,7 +6,7 @@
     <title>Fling Player</title>
 
     <!-- stylesheets of building blocks -->
-    <link rel="stylesheet" type="text/css" href="tv_shared/font/gaia-device-icons.css">
+    <link rel="stylesheet" type="text/css" href="bower_component/smart-icons/smart-icons.css">
     <link rel="stylesheet" type="text/css" href="bower_components/smart-button/style.css">
     <link rel="stylesheet" type="text/css" href="style/fling_player.css">
