+++ tv_apps/smart-home/bower.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "evt": "gaia-components/evt#~0.4.0",
     "smart-button": "smart-components/smart-button",
-    "menu-group": "smart-components/menu-group"
+    "menu-group": "smart-components/menu-group",
+    "smart-icons": "smart-components/smart-icons"
   }
 }
