+++ tv_apps/smart-home/index.html
@@ -6,26 +6,26 @@
     <title>Home</title>
 
     <!-- building blocks -->
-    <link rel="stylesheet" type="text/css" href="tv_shared/font/gaia-device-icons.css">
+    <link rel="stylesheet" type="text/css" href="bower_components/smart-icons/smart-icons.css">
     <link rel="stylesheet" type="text/css" href="bower_components/smart-button/style.css">
     <link rel="stylesheet" type="text/css" href="bower_components/menu-group/style.css">
-    <link rel="stylesheet" type="text/css" href="tv_shared/style/animations.css">
+    <link rel="stylesheet" type="text/css" href="shared/style/smart-screen/animations.css">
 
     <!-- home -->
     <link rel="stylesheet" type="text/css" href="style/header.css">
     <link rel="stylesheet" type="text/css" href="style/home.css">
     <link rel="stylesheet" type="text/css" href="style/deck.css">
     <link rel="stylesheet" type="text/css" href="style/edit.css">
-    <link rel="stylesheet" type="text/css" href="tv_shared/style/navigation.css">
+    <link rel="stylesheet" type="text/css" href="shared/style/smart-screen/navigation.css">
     <link rel="stylesheet" type="text/css" href="style/info-bar.css">
     <link rel="stylesheet" type="text/css" href="style/search-bar.css">
 
     <!-- default cards for first time launch -->
-    <link rel="resource" type="application/json" href="tv_shared/resources/default-cards.json">
+    <link rel="resource" type="application/json" href="shared/resources/default-cards.json">
 
     <!-- For test purpose on phone only -->
     <script defer src="shared/js/gesture_detector.js"></script>
-    <script defer src="tv_shared/js/init_gesture.js"></script>
+    <script defer src="shared/js/smart-screen/init_gesture.js"></script>
 
     <!-- Shared libraries -->
     <script defer src="shared/js/uuid.js"></script>
@@ -33,17 +33,17 @@
 
     <!-- Helper libraries -->
     <script defer src="bower_components/evt/index.js"></script>
-    <script defer src="tv_shared/js/shared_utils.js"></script>
-    <script defer src="tv_shared/js/clock.js"></script>
-    <script defer src="tv_shared/js/piped_promise.js"></script>
-    <script defer src="tv_shared/js/card_store.js"></script>
-    <script defer src="tv_shared/js/cards/card.js"></script>
-    <script defer src="tv_shared/js/cards/application.js"></script>
-    <script defer src="tv_shared/js/cards/deck.js"></script>
-    <script defer src="tv_shared/js/cards/folder.js"></script>
-    <script defer src="tv_shared/js/async_semaphore.js"></script>
-    <script defer src="tv_shared/js/card_manager.js"></script>
-    <script defer src="tv_shared/js/animations.js"></script>
+    <script defer src="shared/js/smart-screen/shared_utils.js"></script>
+    <script defer src="shared/js/smart-screen/clock.js"></script>
+    <script defer src="shared/js/smart-screen/piped_promise.js"></script>
+    <script defer src="shared/js/smart-screen/card_store.js"></script>
+    <script defer src="shared/js/smart-screen/cards/card.js"></script>
+    <script defer src="shared/js/smart-screen/cards/application.js"></script>
+    <script defer src="shared/js/smart-screen/cards/deck.js"></script>
+    <script defer src="shared/js/smart-screen/cards/folder.js"></script>
+    <script defer src="shared/js/smart-screen/async_semaphore.js"></script>
+    <script defer src="shared/js/smart-screen/card_manager.js"></script>
+    <script defer src="shared/js/smart-screen/animations.js"></script>
 
     <!-- building blocks -->
     <script defer src="bower_components/smart-button/script.js"></script>
@@ -51,9 +51,9 @@
     <script defer src="bower_components/menu-group/script.js"></script>
 
     <!-- Shared TV library for keyboard-based navigating -->
-    <script defer src="tv_shared/js/scrollable.js"></script>
-    <script defer src="tv_shared/js/spatial_navigator.js"></script>
-    <script defer src="tv_shared/js/key_navigation_adapter.js"></script>
+    <script defer src="shared/js/smart-screen/scrollable.js"></script>
+    <script defer src="shared/js/smart-screen/spatial_navigator.js"></script>
+    <script defer src="shared/js/smart-screen/key_navigation_adapter.js"></script>
     <!-- Specific code -->
     <script defer src="js/card_filter.js"></script>
     <script defer src="js/connection_manager.js"></script>
