+++ tv_apps/smart-settings/index.html
@@ -5,7 +5,7 @@
     <meta charset="utf-8">
     <title>Home</title>
 
-    <link rel="stylesheet" type="text/css" href="tv_shared/style/navigation.css">
+    <link rel="stylesheet" type="text/css" href="shared/style/smart-screen/navigation.css">
     <link rel="stylesheet" type="text/css" href="style/gaia-icons.css">
     <link rel="stylesheet" type="text/css" href="style/settings.css">
     <link rel="stylesheet" type="text/css" href="style/menu-items.css">
@@ -17,10 +17,10 @@
 
     <!-- Helper libraries -->
     <script defer src="bower_components/evt/index.js"></script>
-    <script defer src="tv_shared/js/selection_border.js"></script>
-    <script defer src="tv_shared/js/spatial_navigator.js"></script>
-    <script defer src="tv_shared/js/applications.js"></script>
-    <script defer src="tv_shared/js/settings_cache.js"></script>
+    <script defer src="shared/js/smart-screen/selection_border.js"></script>
+    <script defer src="shared/js/smart-screen/spatial_navigator.js"></script>
+    <script defer src="shared/js/smart-screen/applications.js"></script>
+    <script defer src="shared/js/smart-screen/settings_cache.js"></script>
 
     <!-- main -->
     <script defer src="js/base.js"></script>
