+++ tv_apps/smart-system/index.html
@@ -20,12 +20,12 @@
     <script defer src="shared/js/notification_helper.js"></script>
     <!-- We need this before we can show the boot animation -->
     <script src="shared/js/async_storage.js"></script>
-    <script src="tv_shared/js/settings_cache.js"></script>
-    <script src="tv_shared/js/animations.js"></script>
+    <script src="shared/js/smart-screen/settings_cache.js"></script>
+    <script src="shared/js/smart-screen/animations.js"></script>
     <!-- focus manager -->
     <script src="js/focus_manager.js"></script>
 
-    <script defer src="tv_shared/js/shared_utils.js"></script>
+    <script defer src="shared/js/smart-screen/shared_utils.js"></script>
     <script defer src="bower_components/evt/index.js"></script>
     <script defer src="shared/js/manifest_helper.js"></script>
     <script defer src="shared/js/mime_mapper.js"></script>
@@ -38,8 +38,8 @@
     <script defer src="shared/js/font_size_utils.js"></script>
     <!-- XXX can be removed while gecko support navigator.mozHour12 API -->
     <script defer src="shared/js/date_time_helper.js"></script>
-    <script defer src="tv_shared/js/simple_key_navigation.js"></script>
-    <script defer src="tv_shared/js/key_navigation_adapter.js"></script>
+    <script defer src="shared/js/smart-screen/simple_key_navigation.js"></script>
+    <script defer src="shared/js/smart-screen/key_navigation_adapter.js"></script>
 
     <script defer src="js/service.js"></script>
     <script defer src="js/base_ui.js"></script>
@@ -63,8 +63,8 @@
 
     <link rel="stylesheet" type="text/css" href="style/definitions.css">
     <link rel="stylesheet" type="text/css" href="style/attention_indicator.css">
-    <link rel="stylesheet" type="text/css" href="tv_shared/style/navigation.css">
-    <link rel="stylesheet" type="text/css" href="tv_shared/style/animations.css">
+    <link rel="stylesheet" type="text/css" href="shared/style/smart-screen/navigation.css">
+    <link rel="stylesheet" type="text/css" href="shared/style/smart-screen/animations.css">
 
     <!-- System -->
     <link rel="stylesheet" type="text/css" href="style/system/system.css">
@@ -126,7 +126,7 @@
     <script defer src="js/external_storage_monitor.js"></script>
     <script defer src="js/bootstrap.js"></script>
     <script defer src="js/screenshot.js"></script>
-    <script defer src="tv_shared/js/async_semaphore.js"></script>
+    <script defer src="shared/js/smart-screen/async_semaphore.js"></script>
     <script defer src="js/sound_manager.js"></script>
     <script defer src="js/browser_key_event_manager.js"></script>
     <script defer src="js/hardware_buttons.js"></script>
@@ -269,7 +269,7 @@
     <link rel="stylesheet" type="text/css" href="style/app_titlebar.css">
     <link rel="stylesheet" type="text/css" href="style/browser_context_menu/browser_context_menu.css">
     <script defer src="js/base_ui.js"></script>
-    <script defer src="tv_shared/js/scrollable.js"></script>
+    <script defer src="shared/js/smart-screen/scrollable.js"></script>
     <script defer src="js/browser_context_menu.js"></script>
     <script defer src="js/child_window_factory.js"></script>
     <script defer src="js/app_modal_dialog.js"></script>
@@ -320,8 +320,8 @@
     <!-- <script defer src="shared/js/icons_helper.js"></script> -->
 
     <!-- Navigation -->
-    <script defer src="tv_shared/js/spatial_navigator.js"></script>
-    <script defer src="tv_shared/js/init_gesture.js"></script>
+    <script defer src="shared/js/smart-screen/spatial_navigator.js"></script>
+    <script defer src="shared/js/smart-screen/init_gesture.js"></script>
 
     <!-- API Services -->
     <script defer src="js/screenshot_service.js"></script>
